Made Password writable in BeforeInstall
I want to modify the Password in the BeforeInstall event.
Just making it writable as suggested did not work because
HostServiceInstaller is created with command line passed Password
already and then only the BeforeInstall is called. So it is too late to
just set the new Password in BeforeInstall?

See https://groups.google.com/forum/#!topic/topshelf-discuss/4iBCSS0WW4c