Add Support for Watching Log4Net Files
Log4net can be configured to watch the config file for changes and
re-load it witout stopping the applicaiton. This is quite useful for
services.

You can configure log4net to do this with assembly attributes or an
entry in the App.config file. This adds the ability to specify it in the
`.UseLog4Net` call too for ortogonality.